refactor: progress bar component/common으로 이동
