+++ src/component/common/ProgressBar/ProgressBar.module.scss
