+++ src/page/Auth/Signup/index.tsx
@@ -3,8 +3,8 @@ import useMediaQuery from 'utils/hooks/useMediaQuery';
 import { ReactComponent as Logo } from 'assets/svg/auth/koin-logo.svg';
 import { ReactComponent as Back } from 'assets/svg/common/back-arrow.svg';
 import { Link } from 'react-router-dom';
+import ProgressBar from 'component/common/ProgressBar';
 import UserEmail from './component/UserEmail';
-import ProgressBar from './component/ProgressBar';
 import OwnerData from './view/OwnerDataPage';
 import TermsOfService from './view/TermsOfServicePage';
 import UserData from './view/UserDataPage';
